requirements.txt
beautifulsoup4>=4.12.0
httpx>=0.27.0
openai>=1.30.0
pydantic>=2.6.0
python-dotenv>=1.0.0
